+++ components/ocs_system/log_updater.h
@@ -23,7 +23,7 @@ public:
     LogUpdater(IUpdater& updater, const char* log_tag);
 
     //! Begin firmware update.
-    status::StatusCode begin(size_t total_size, uint32_t crc32) override;
+    status::StatusCode begin(uint32_t total_size, uint32_t crc32) override;
 
     //! Write firmware data.
     status::StatusCode write(const uint8_t* buf, size_t len) override;
