+++ components/ocs_system/target_esp32/spiffs_updater.h
@@ -23,7 +23,7 @@ public:
     SpiffsUpdater(IUpdater& updater, const char* partition_label);
 
     //! Unmount SPIFFS partition and begin the firmware update process.
-    status::StatusCode begin(size_t total_size, uint32_t crc32) override;
+    status::StatusCode begin(uint32_t total_size, uint32_t crc32) override;
 
     //! Write firmware data.
     status::StatusCode write(const uint8_t* buf, size_t len) override;
